Fix asset enqueuing
Enqueue assets on the front end. It looks like this bug was introduced
in the 2.0 release. Prior to the version bump, [assets were being
enqueued] from the `admin_init` and `init` hooks. However, after the
update, [front end assets were no longer being enqueued].

This commit reimplements the original behavior so that assets are
enqueued on the front end. Additionally, this commit prefixes style and
script handles with `tsl`, and moves a bit of JavaScript out of a
template and into a dedicated file. In general, I tried to make
everything related to asset enqueuing as conventional and as possible.

[assets were being enqueued]: https://github.com/seagyn/the-social-links/blob/a3c4136e5f6294d886882ecde463ae996be6a838/the-social-links.php#L84-L85
[front end assets were no longer being enqueued]: https://github.com/seagyn/the-social-links/blob/afd7197c44a079af88c72cc82f58f56a4f2fdc63/src/boot.php#L49